requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-extensions==3.2.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
